+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-
+- profile: show local time of cursor position in tooltip (fixes #3469).
 ---
 * Always add new entries at the very top of this file above other existing entries and this note.
 * Use this layout for new entries: `[Area]: [Details about the change] [reference thread / issue]`